requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.5.0
numpy
Pillow
scipy>=0.17.0
